Move config.py to src/python
