+++ docker-compose.override.yaml
@@ -4,7 +4,7 @@ services:
     ports:
       - "5000:5000"
     volumes:
-      - ./src/python/api:/usr/src/app
+      - ./src/python:/usr/src/app
     environment:
       - PORT=5000
       - FLASK_DEBUG=True
@@ -20,7 +20,7 @@ services:
       - FLASK_SECRET_KEY=testing
   migration:
     volumes:
-      - ./src/python/api:/usr/src/app
+      - ./src/python:/usr/src/app
     environment:
       - FLASK_DEBUG=True
       - DB_HOST=pg
