+++ src/python/api/app.py
@@ -9,7 +9,7 @@ from sqlalchemy.dialects.postgresql import insert
 from typing import Optional
 from urllib.parse import urlencode
 
-from .config import app, db, verify_fitbit_signature, verify_fitbit_verification
+from ..config import app, db, verify_fitbit_signature, verify_fitbit_verification
 from . import models
 from . import gql
