+++ src/python/api/gql/types/fitbit_authorization.py
@@ -10,7 +10,7 @@ from graphql import (
     GraphQLString,
 )
 from urllib.parse import urlencode
-from ...config import app
+from ....config import app
 
 
 def fitbit_authorization_fields() -> dict[str, GraphQLField]:
